// apps/web/src/types/company.ts
export interface Company {
  id: string;
  name: string;
  address?: string;
  tin?: string;
  isActive: boolean;
  isDeleted?: boolean;
  deletedAt?: Date | null;
  printHeader?: string;
  printFooter?: string;
  printCss?: string;
  defaultWorkdays?: number;
  currency?: string;
  payrollPeriods?: Array<{
    type: "monthly" | "semi-monthly" | "bi-weekly" | "weekly";
    cutOff1Day?: number;
    cutOff2Day?: number;
    payDay?: number;
    frequency?: string;
  }>;
  columnGroup?: {
    dtr: boolean;
    salaries: boolean;
    earnings: boolean;
    benefits: boolean;
    deductions: boolean;
  };
  createdAt: Date;
  updatedAt: Date;
}

export interface CompanyPeriod {
  id: string;
  companyId: string;
  type: "monthly" | "semi-monthly" | "weekly" | "bi-weekly";
  payDay: number;
}

export interface CompanyOptions {
  id: string;
  companyId: string;
  columnGroup: {
    dtr: boolean;
    salaries: boolean;
    earnings: boolean;
    benefits: boolean;
    deductions: boolean;
  };
  workDays: number[];
  printGroup?: string;
  payslipTemplate?: string;
}